+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Fix skipping partial withdrawals count.
 - recover from panics when writing the event stream [pr](https://github.com/prysmaticlabs/prysm/pull/14545)
 - Return the correct light client payload proof. [PR](https://github.com/prysmaticlabs/prysm/pull/14565)
+- Set fields of wrapped proto object in light client setters. [PR](https://github.com/prysmaticlabs/prysm/pull/14573)
 
 ### Security